+++ pages/2-material/styles.scss
@@ -1,0 +1,4 @@
+.material {
+  background-color: rgb(255, 255, 255);
+  padding: 40px;
+}
