+++ pages/3-aura/styles.scss
@@ -1,3 +1,8 @@
+.aura {
+  background-color: rgb(255, 255, 255);
+  padding: 40px;
+}
+
 .aura-checkbox {
   background-color: #fff;
   border-radius: 8px;
