+++ pages/4-synthesis/styles.scss
@@ -1,0 +1,4 @@
+.synthesis {
+  background-color: rgb(255, 255, 255);
+  padding: 40px;
+}
