+++ pages/5-katal/styles.scss
@@ -1,0 +1,4 @@
+.katal {
+  background-color: rgb(255, 255, 255);
+  padding: 40px;
+}
